chore: move event listener to room package
